Remove duplicate store
